# Use inheritance instead of interfaces