--- a/samples/Esprima.Benchmark/AstTreeWalkBenchmark.cs
+++ b/samples/Esprima.Benchmark/AstTreeWalkBenchmark.cs
@@ -24,9 +24,8 @@ namespace Esprima.Benchmark
             var walker = new ScriptWalker();
             walker.Visit(_script);
             if (walker._lexicalNameCount != 0) throw new InvalidOperationException("wrong _lexicalNameCount" + walker._lexicalNameCount);
-            if (walker._varNameCount != 1849) throw new InvalidOperationException("wrong _varNameCount " + walker._varNameCount);
+            if (walker._varNameCount != 1856) throw new InvalidOperationException("wrong _varNameCount " + walker._varNameCount);
             if (walker._functionCount != 1610) throw new InvalidOperationException("wrong _functionCount " + walker._functionCount);
-            if (walker._visitCount != 32315) throw new InvalidOperationException("wrong _visitCount " + walker._visitCount);
         }
 
         private sealed class ScriptWalker