--- a/src/Esprima/Ast/ArgumentListElement.cs
+++ b/src/Esprima/Ast/ArgumentListElement.cs
@@ -1,6 +1,0 @@
-﻿namespace Esprima.Ast
-{
-    public interface ArgumentListElement : INode
-    {
-    }
-}