--- a/src/Esprima/Ast/ArrayExpression.cs
+++ b/src/Esprima/Ast/ArrayExpression.cs
@@ -2,19 +2,19 @@ using System.Collections.Generic;
 
 namespace Esprima.Ast
 {
-    public class ArrayExpression : Node, Expression
+    public class ArrayExpression : Expression
     {
-        private readonly NodeList<ArrayExpressionElement> _elements;
+        private readonly NodeList<Expression> _elements;
 
-        public ArrayExpression(in NodeList<ArrayExpressionElement> elements) :
+        public ArrayExpression(in NodeList<Expression> elements) :
             base(Nodes.ArrayExpression)
         {
             _elements = elements;
         }
 
-        public ref readonly NodeList<ArrayExpressionElement> Elements => ref _elements;
+        public ref readonly NodeList<Expression> Elements => ref _elements;
 
-        public override IEnumerable<INode> ChildNodes =>
+        public override IEnumerable<Node> ChildNodes =>
             ChildNodeYielder.Yield(_elements);
     }
 }