--- a/src/Esprima/Ast/ArrayExpression.cs
+++ b/src/Esprima/Ast/ArrayExpression.cs
@@ -2,19 +2,17 @@ using System.Collections.Generic;
 
 namespace Esprima.Ast
 {
-    public class ArrayExpression : Expression
+    public sealed class ArrayExpression : Expression
     {
         private readonly NodeList<Expression> _elements;
 
-        public ArrayExpression(in NodeList<Expression> elements) :
-            base(Nodes.ArrayExpression)
+        public ArrayExpression(in NodeList<Expression> elements) : base(Nodes.ArrayExpression)
         {
             _elements = elements;
         }
 
         public ref readonly NodeList<Expression> Elements => ref _elements;
 
-        public override IEnumerable<Node> ChildNodes =>
-            ChildNodeYielder.Yield(_elements);
+        public override IEnumerable<Node> ChildNodes => ChildNodeYielder.Yield(_elements);
     }
 }