--- a/src/Esprima/Ast/ArrayExpressionElement.cs
+++ b/src/Esprima/Ast/ArrayExpressionElement.cs
@@ -1,6 +1,0 @@
-﻿namespace Esprima.Ast
-{
-    public interface ArrayExpressionElement : INode
-    {
-    }
-}