--- a/src/Esprima/Ast/ArrayPattern.cs
+++ b/src/Esprima/Ast/ArrayPattern.cs
@@ -11,6 +11,6 @@
 
         public ref readonly NodeList<Expression> Elements => ref _elements;
 
-        public override NodeCollection ChildNodes => ChildNodeYielder.Yield(_elements);
+        public override NodeCollection ChildNodes => GenericChildNodeYield.Yield(_elements);
     }
 }