--- a/src/Esprima/Ast/ArrayPatternElement.cs
+++ b/src/Esprima/Ast/ArrayPatternElement.cs
@@ -1,0 +1,9 @@
+﻿namespace Esprima.Ast
+{
+    public abstract class ArrayPatternElement : Expression
+    {
+        protected ArrayPatternElement(Nodes type) : base(type)
+        {
+        }
+    }
+}