--- a/src/Esprima/Ast/AssignmentExpression.cs
+++ b/src/Esprima/Ast/AssignmentExpression.cs
@@ -34,19 +34,18 @@ namespace Esprima.Ast
         ExponentiationAssign,
     }
 
-    public class AssignmentExpression : Node,
-        Expression
+    public class AssignmentExpression : Expression
     {
         public readonly AssignmentOperator Operator;
 
         // Can be something else than Expression (ObjectPattern, ArrayPattern) in case of destructuring assignment
-        public readonly INode Left;
+        public readonly Expression Left;
         public readonly Expression Right;
 
-        public AssignmentExpression(string op, INode left, Expression right) :
+        public AssignmentExpression(string op, Expression left, Expression right) :
             base(Nodes.AssignmentExpression)
         {
-            Operator = AssignmentExpression.ParseAssignmentOperator(op);
+            Operator = ParseAssignmentOperator(op);
             Left = left;
             Right = right;
         }
@@ -87,7 +86,7 @@ namespace Esprima.Ast
             }
         }
 
-        public override IEnumerable<INode> ChildNodes =>
+        public override IEnumerable<Node> ChildNodes =>
             ChildNodeYielder.Yield(Left, Right);
     }
 }