--- a/src/Esprima/Ast/AssignmentPattern.cs
+++ b/src/Esprima/Ast/AssignmentPattern.cs
@@ -2,19 +2,17 @@
 
 namespace Esprima.Ast
 {
-    public class AssignmentPattern : ArrayPatternElement
+    public sealed class AssignmentPattern : ArrayPatternElement
     {
         public readonly Node Left;
         public Node Right;
 
-        public AssignmentPattern(Node left, Node right) :
-            base(Nodes.AssignmentPattern)
+        public AssignmentPattern(Node left, Node right) : base(Nodes.AssignmentPattern)
         {
             Left = left;
             Right = right;
         }
 
-        public override IEnumerable<Node> ChildNodes =>
-            ChildNodeYielder.Yield(Left, Right);
+        public override IEnumerable<Node> ChildNodes => ChildNodeYielder.Yield(Left, Right);
     }
 }