--- a/src/Esprima/Ast/AwaitExpression.cs
+++ b/src/Esprima/Ast/AwaitExpression.cs
@@ -9,6 +9,6 @@ namespace Esprima.Ast
             Argument = argument;
         }
 
-        public override NodeCollection ChildNodes => ChildNodeYielder.Yield(Argument);
+        public override NodeCollection ChildNodes => new NodeCollection(Argument);
     }
 }