--- a/src/Esprima/Ast/BinaryExpression.cs
+++ b/src/Esprima/Ast/BinaryExpression.cs
@@ -56,7 +56,7 @@ namespace Esprima.Ast
         Exponentiation,
     }
 
-    public class BinaryExpression : Expression
+    public sealed class BinaryExpression : Expression
     {
         public readonly BinaryOperator Operator;
         public readonly Expression Left;