--- a/src/Esprima/Ast/BinaryExpression.cs
+++ b/src/Esprima/Ast/BinaryExpression.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Collections.Generic;
 using Esprima.Utils;
 
 namespace Esprima.Ast
@@ -130,7 +129,7 @@ namespace Esprima.Ast
             }
         }
 
-        public override IEnumerable<Node> ChildNodes =>
+        public override NodeCollection ChildNodes =>
             ChildNodeYielder.Yield(Left, Right);
     }
 }