So now that Jint is going to do some heave traversal of AST I decided to look into improving performance in that area. All these empty interfaces have always bothered me and they are never used (even in Jint that does handle almost the full AST). I think the origin has been TypeScript typing in Esprima side where the rules have made sense as TS has all kinds of union types to use, however c# doesn't.

In this PR:

* `INode` is removed, it's basically `Node`, it's faster to access, no interface dispatch
* Removed all interfaces except for IFunction, which ties function expressions and declarations
* everything is basically an `Expression` or `Statement` (which both are `StatementListItem`s)
* sealing AST types
* less allocations in enumeration of children, now using concrete struct enumerator type
* also tweaked NodeList and ArrayList's enumerator to behave like List.Enumerator in BCL

I've already found some logic errors in Jint (non-critical) but the interfaces hid some code paths that were never accessed.