@sebastienros 2020-05-07
Much simpler to read obviously. I see some types got removed, at first I wondered how you would, but they are just conveying type information, and mostly useless. I probably used them to follow the esprima semantics. LGTM.