@lahma 2020-05-09
Now performance regressions. AST walk is slightly slower due to enumerator logic but now basically allocation free.

## Esprima.Benchmark.AstTreeWalkBenchmark

| **Diff**|Method|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |VisitChildren|5.588 ms|0.0102 ms|773.4375|0.0000|0.0000|3.09 MB|
| **New** |	| **6.178 ms (+11%)** | **0.0734 ms** | **0.0000 (-100%)** | **0.0000** | **0.0000** | **42 B (-100%)** |


## Esprima.Benchmark.FileParsingBenchmark

| **Diff**|Method|FileName|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |ParseProgram|angular-1.2.5|28.292 ms|0.2144 ms|2343.7500|968.7500|125.0000|13.11 MB|
| **New** |	|	| **27.797 ms (-2%)** | **0.3874 ms** | **2343.7500 (0%)** | **968.7500 (0%)** | **125.0000 (0%)** | **13.11 MB (0%)** |
| Old |ParseProgram|backbone-1.1.0|3.196 ms|0.0158 ms|378.9063|187.5000|0.0000|2.07 MB|
| **New** |	|	| **3.094 ms (-3%)** | **0.0281 ms** | **378.9063 (0%)** | **187.5000 (0%)** | **0.0000** | **2.07 MB (0%)** |
| Old |ParseProgram|jquery-1.9.1|22.534 ms|0.3415 ms|2031.2500|968.7500|125.0000|11.43 MB|
| **New** |	|	| **22.282 ms (-1%)** | **0.3536 ms** | **2031.2500 (0%)** | **968.7500 (0%)** | **125.0000 (0%)** | **11.43 MB (0%)** |
| Old |ParseProgram|jquery.mobile-1.4.2|40.064 ms|0.4953 ms|3230.7692|923.0769|307.6923|17.7 MB|
| **New** |	|	| **39.310 ms (-2%)** | **0.4558 ms** | **3230.7692 (0%)** | **923.0769 (0%)** | **307.6923 (0%)** | **17.7 MB (0%)** |
| Old |ParseProgram|mootools-1.4.5|18.690 ms|0.3704 ms|1718.7500|843.7500|125.0000|9.36 MB|
| **New** |	|	| **18.285 ms (-2%)** | **0.3598 ms** | **1781.2500 (+4%)** | **906.2500 (+7%)** | **156.2500 (+25%)** | **9.36 MB (0%)** |
| Old |ParseProgram|underscore-1.5.2|2.590 ms|0.0095 ms|359.3750|136.7188|0.0000|1.8 MB|
| **New** |	|	| **2.562 ms (-1%)** | **0.0143 ms** | **359.3750 (0%)** | **136.7188 (0%)** | **0.0000** | **1.8 MB (0%)** |
| Old |ParseProgram|yui-3.12.0|16.513 ms|0.3753 ms|1593.7500|781.2500|93.7500|8.56 MB|
| **New** |	|	| **16.383 ms (-1%)** | **0.2377 ms** | **1593.7500 (0%)** | **781.2500 (0%)** | **93.7500 (0%)** | **8.56 MB (0%)** |


## Jint.Benchmark.EvaluationBenchmark

| **Diff**|Method|N|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |ParseProgram|200|5.583 ms|0.0312 ms|1265.6250|0.0000|0.0000|5.07 MB|
| **New** |	|	| **5.550 ms (-1%)** | **0.0438 ms** | **1265.6250 (0%)** | **0.0000** | **0.0000** | **5.07 MB (0%)** |