@sebastienros 2020-05-12
Bump https://github.com/sebastienros/esprima-dotnet/blob/dev/src/Esprima/Esprima.csproj#L17 to 2.0.x ?

Then I would create a release/1.x branch to track the previous version, master would then become 2.0.x

Also update the benchmarks table on the readme?