@lahma 2020-05-12
Did the version change. How did you get the perf numbers for original Esprima?